Move hooks into its own folder
